Simulate async add project operation
If the add project operation were backed by a network call, the latency would
be nondeterministic. We can simulate this by scheduling on a new scheduler and
sleeping for a random amount of time. Because projects might be added in any
order, I updated the assertion to compare sets instead of arrays.