Simplify `-[id<APIClient> projects]` signal
My original idea was that this signal would return a signal of arrays of
signals. This commit changes this to a signal of signals, which is much simpler
to work with.

The ReactiveCocoa library has methods like `-[RACSignal flatten]` specifically
for this type of signal.

Consumers can use `-[RACSignal bufferWithTime:onScheduler]` to transform the
signal back into a signal of collections of signals.